spot has *piece move piece(..) sets that and pays if passed go

+++ MonopolyProject/Piece.h
@@ -6,7 +6,7 @@ public:
 	Piece();
 	Piece(std::string name);
 	std::string name;
-	int position;
+	unsigned short position;
 	bool operator==(const Piece& other);
 };
